Another approach for the program using BK-Trees to
reduce distances computations. The idea is to load the dictionary
into one of these trees, however this is showing to be a heavy
task, taking up a few seconds in contrast with the trie

Also, this may work with smaller dictionaries, as tests with the usual
dictionary take up to 3.5GB of memory during the program.