Added preprocessor flag for conditional compiling.
This flag determines whether the program will check for duplicates
when generating suggestions. In practice, this makes the program
a little slower for low search distances.

Updated Makefile to accept this compiler flag.

Directory dictionaries contains test dictionaries, the program now
asks from standard input for a dictionary path.

Since strdup is not in some gcc versions of the c99 standard, copy_str
was added again in utils.c